Fixes import after rename
