+++ services/storage/src/simcore_service_storage/models.py
@@ -4,9 +4,8 @@
 from typing import Tuple
 
 import attr
-
-from simcore_postgres_database.storage_tables import (file_meta_data, metadata,
-                                                    tokens)
+from simcore_postgres_database.storage_models import (file_meta_data, metadata,
+                                                      tokens)
 
 from .settings import DATCORE_STR, SIMCORE_S3_ID, SIMCORE_S3_STR
